requirements.txt
transformers>=4.30.0
datasets>=2.12.0
torch>=2.0.0
python-dotenv>=1.0.0
azure-cognitiveservices-vision-customvision>=3.1.0
pillow>=9.5.0
numpy>=1.24.0
requests>=2.31.0
scikit-learn>=1.3.0
pandas>=2.0.0
kagglehub>=0.1.0
ultralytics>=8.0.0  # For YOLOv8
tqdm>=4.65.0
matplotlib>=3.7.0
opencv-python>=4.8.0